Homework Spring-Boot_Add-Tests-Books-Categories. Added Optional.empty() and created folder structure.
